turning fixed i think,

+++ Software/STM32/thikc/Inc/pid.h
@@ -4,14 +4,14 @@
 #define pid
 
 typedef struct{
-    uint8_t kp;
-    uint8_t kd;
-    uint8_t ki;
-    int16_t i;
-    int8_t i_max;
+    uint16_t kp;
+    uint16_t kd;
+    uint16_t ki;
+    int32_t i;
+    int16_t i_max;
     uint16_t gain;
     int32_t err;
-    int16_t count;
+    uint32_t count;
 } PID_t;
 
 PID_t pid_Init(uint16_t kp, uint16_t kd, uint16_t ki, int16_t i_max, uint16_t g);
